backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.43
aiosqlite==0.21.0
python-dotenv==1.2.1
google-generativeai==0.3.2
duckduckgo-search==8.1.1
